requirements.txt
Django==1.8.3
djangorestframework==3.1.3
djoser==0.3.1
ipython==3.2.1
Markdown==2.6.2
